Refactoring sub-command directory structure
