+++ analyzer/codechecker_analyzer/cmd/analyzer_version.py
