+++ analyzer/codechecker_analyzer/cmd/analyzers.py
