+++ analyzer/codechecker_analyzer/cmd/checkers.py
