+++ analyzer/codechecker_analyzer/cmd/log.py
