+++ analyzer/tools/build-logger/README.md
@@ -7,7 +7,6 @@ This tool can capture the build process and generate a
 
 To build the project execute
 ~~~~~~~
-cd analyzer/tools/build-logger
 make -f Makefile.manual
 ~~~~~~~
