+++ libcodechecker/cmd/__init__.py
