+++ web/client/codechecker_client/cmd/cmd.py
