+++ web/client/codechecker_client/cmd/store.py
